Removed Kickstart and started porting my own Neovim config to Lua
I didn't like the folder structure for plugins in Kickstart and wanted
to remove a bunch of plugins, but ran into some errors that I couldn't
fix. So I figured it's better to start from scratch, so I know what I'm
doing. Still the Kickstart configuration is very useful as a reference.